# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  planning.Epoch:
    properties:
      createdAt:
        type: string
      daysRemaining:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      name:
        type: string
      phase:
        type: string
      projectId:
        type: integer
      startDate:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  planning.Goal:
    properties:
      createdAt:
        type: string
      description:
        type: string
      epochId:
        type: integer
      id:
        type: integer
      ownerUserId:
        type: integer
      progressPercent:
        type: integer
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  planning.deleteResponse:
    properties:
      deleted:
        example: true
        type: boolean
    type: object
  planning.epochRequest:
    properties:
      endDate:
        type: string
      name:
        type: string
      phase:
        type: string
      projectId:
        type: integer
      startDate:
        type: string
      status:
        type: string
    type: object
  planning.errorResponse:
    properties:
      code:
        example: bad_request
        type: string
      message:
        example: validation failed
        type: string
    type: object
  planning.goalRequest:
    properties:
      description:
        type: string
      epochId:
        type: integer
      ownerUserId:
        type: integer
      progressPercent:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Seamless collaboration backend.
  title: Seamless Service API
  version: 0.2.0
paths:
...
  /epochs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/planning.Epoch'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: List epochs
      tags:
      - Planning
    post:
      consumes:
      - application/json
      parameters:
      - description: Epoch payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/planning.epochRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/planning.Epoch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Create epoch
      tags:
      - Planning
  /epochs/{epochId}:
    delete:
      parameters:
      - description: Epoch ID
        in: path
        name: epochId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/planning.deleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Delete epoch
      tags:
      - Planning
    get:
      parameters:
      - description: Epoch ID
        in: path
        name: epochId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/planning.Epoch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Get epoch
      tags:
      - Planning
    patch:
      consumes:
      - application/json
      parameters:
      - description: Epoch ID
        in: path
        name: epochId
        required: true
        type: integer
      - description: Epoch payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/planning.epochRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/planning.Epoch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Update epoch
      tags:
      - Planning
  /goals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/planning.Goal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: List goals
      tags:
      - Planning
    post:
      consumes:
      - application/json
      parameters:
      - description: Goal payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/planning.goalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/planning.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Create goal
      tags:
      - Planning
  /goals/{goalId}:
    delete:
      parameters:
      - description: Goal ID
        in: path
        name: goalId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/planning.deleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Delete goal
      tags:
      - Planning
    get:
      parameters:
      - description: Goal ID
        in: path
        name: goalId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/planning.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Get goal
      tags:
      - Planning
    patch:
      consumes:
      - application/json
      parameters:
      - description: Goal ID
        in: path
        name: goalId
        required: true
        type: integer
      - description: Goal payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/planning.goalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/planning.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/planning.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/planning.errorResponse'
      security:
      - bearerAuth: []
      summary: Update goal
      tags:
      - Planning
...
schemes:
- http
- https
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"